+++ HISTORY.md
@@ -1,3 +1,7 @@
+# v1.0.2
+* Model update to `UpdateCheckout.RequestedInfo` to clear "email" and "name" flags.
+* Added more XMLDoc comments. 
+
 # v1.0.1
 * Added `.HasError()` and `.HasWarnings()` helper methods on JSON response objects.
 * Set `X-CC-Version` header to `2018-03-22` to remove server warnings.